requirements.txt
Flask
python-docx
PyMuPDF
Werkzeug
